feat: parse "LONGDATE" field
The specification describes the purpose of the LONGDATE field as
"Date and time when spectrum was measured" (see the previous "DATE" and
"TIME" fields).

The specification of the field's format is a bit ambiguous regarding the
sub-second token.
Its length (number of digits) and meaning (ms/us/ns?) is not defined.
This implementation guesses the content of the token based on the number
of parsed digits.

In case of a parsing failure (e.g. the LONGDATE field contains an
invalid string), then the original string is kept unchanged.